infra: 인프라 클래스 API 모듈로 이동
