+++ Api/src/main/java/com/mapz/api/common/image/s3/S3Api.java
@@ -1,4 +1,4 @@
-package com.cheocharm.MapZ.common.image.s3;
+package com.mapz.api.common.image.s3;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
