+++ Api/src/main/java/com/mapz/api/common/log/exception/ExceptionLogSchema.java
@@ -1,8 +1,8 @@
-package com.cheocharm.MapZ.common.log.response;
+package com.mapz.api.common.log.exception;
 
-import com.cheocharm.MapZ.common.CommonResponse;
-import com.cheocharm.MapZ.common.exception.common.FailConvertToJsonException;
-import com.cheocharm.MapZ.common.util.ObjectMapperUtils;
+import com.mapz.api.common.CommonResponse;
+import com.mapz.api.common.exception.common.FailConvertToJsonException;
+import com.mapz.api.common.util.ObjectMapperUtils;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -11,17 +11,19 @@ import java.io.IOException;
 
 @Getter
 @Builder
-public class ResponseLogSchema {
+public class ExceptionLogSchema {
     private String api;
     private String httpMethod;
     private int statusCode;
+    private String customCode;
     private String message;
 
-    public static ResponseLogSchema createLogSchema(HttpServletRequest request, CommonResponse response) {
-        return ResponseLogSchema.builder()
+    public static ExceptionLogSchema createLogSchema(HttpServletRequest request, CommonResponse response) {
+        return ExceptionLogSchema.builder()
                 .api(request.getRequestURI())
                 .httpMethod(request.getMethod())
                 .statusCode(response.getStatusCode())
+                .customCode(response.getCustomCode())
                 .message(response.getMessage())
                 .build();
     }
@@ -37,5 +39,4 @@ public class ResponseLogSchema {
 
         return json;
     }
-
 }
