+++ Domain/src/main/java/com/mapz/domain/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.cheocharm.MapZ.common.config;
+package com.mapz.domain.config;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.springframework.context.annotation.Bean;
